moving compotents into comptents folder and extracting Product name and edition and shop now button into seperate file
