+++ src/compotents/ProductCard.tsx
@@ -1,6 +1,7 @@
 import React, { useEffect, useState } from "react";
 import { Transition } from "@headlessui/react";
 import Image from "next/image";
+import ProudctNameAndEdition from "./ProductNameAndEdition";
 
 export default function ProductCard({
   content,
@@ -100,18 +101,7 @@ export default function ProductCard({
               );
             })}
           </div>
-          <div className="relative flex flex-col items-center flex-wrap top-20 font-[440]">
-            <div className="text-sm flex flex-row space-x-2 place-content-center">
-              <div>{content.name}</div>
-              <div className="text-gray-500 content-center flex">
-                {content.edition}
-              </div>
-            </div>
-            <div className="text-sm">{content.price}</div>
-            <button className="relative hover:bg-orange-500 hover:text-white hover:border-orange-500 tracking-wider text-sm font-normal border rounded border-black text- px-5 py-2 top-6 duration-200">
-              SHOP NOW
-            </button>
-          </div>
+          <ProudctNameAndEdition content={content} />
         </div>
       </div>
     </div>
